move server logic into server.py
